Address remaining attendee review feedback
Move the issue-specific mixed-attendee test out of the new-method test module, link attendee parameters to the specific RFC 5545 section, and combine the attendee assignment example into a functional doctest.
AI-assisted: OpenAI Codex implemented the scoped maintainer feedback. OpenAI GPT-5.6 Sol reviewed the final diff and verification results. The contributor reviewed and validated the changes.

AI Prompts:
claude-sonnet-4-6: hi
